+++ src/ingest_validation_tools/check_factory.py
@@ -3,6 +3,7 @@ from string import Template
 from pathlib import Path
 from sys import stderr
 import json
+from typing import List, Callable, Dict, Any, Iterator
 
 import frictionless
 import requests
@@ -10,8 +11,12 @@ import requests
 
 cache_path = Path(__file__).parent / 'url-status-cache.json'
 
+ErrorIterator = Iterator[frictionless.errors.CellError]
+Row = Dict[str, Any]
+Check = Callable[[Row], ErrorIterator]
 
-def make_checks(schema):
+
+def make_checks(schema) -> List[Check]:
     factory = _CheckFactory(schema)
     return [
         factory.make_url_check(),
@@ -26,14 +31,14 @@ class _CheckFactory():
         self.schema = schema
         self._prev_value_run_length = {}
 
-    def _get_constrained_fields(self, constraint):
+    def _get_constrained_fields(self, constraint: str) -> Dict[str, List]:
         c_c = 'custom_constraints'
         return {
             f['name']: f[c_c][constraint] for f in self.schema['fields']
             if c_c in f and constraint in f[c_c]
         }
 
-    def _check_url_status_cache(self, url):
+    def _check_url_status_cache(self, url: str) -> str:
         if not cache_path.exists():
             cache_path.write_text('{}')
         url_status_cache = json.loads(cache_path.read_text())
@@ -52,7 +57,7 @@ class _CheckFactory():
         return url_status_cache[url]
 
     def make_url_check(self, template=Template(
-            'URL returned $status: "$url"')):
+            'URL returned $status: "$url"')) -> Check:
         url_constrained_fields = self._get_constrained_fields('url')
 
         def url_check(row):
@@ -70,7 +75,7 @@ class _CheckFactory():
 
     def make_sequence_limit_check(self, template=Template(
             'there is a run of $run_length sequential items: Limit is $limit. '
-            'If correct, reorder rows.')):
+            'If correct, reorder rows.')) -> Check:
         sequence_limit_fields = self._get_constrained_fields('sequence_limit')
 
         def sequence_limit_check(row):
@@ -114,7 +119,7 @@ class _CheckFactory():
         return sequence_limit_check
 
     def make_units_check(self, template=Template(
-            'Required when $units_for is filled')):
+            'Required when $units_for is filled')) -> Check:
         units_constrained_fields = self._get_constrained_fields('units_for')
 
         def units_check(row):
@@ -127,7 +132,7 @@ class _CheckFactory():
         return units_check
 
     def make_forbid_na_check(self, template=Template(
-            '"N/A" fields should just be left empty')):
+            '"N/A" fields should just be left empty')) -> Check:
         forbid_na_constrained_fields = self._get_constrained_fields('forbid_na')
 
         def forbid_na_check(row):