+++ src/validate_tsv.py
@@ -8,7 +8,7 @@ from ingest_validation_tools.error_report import ErrorReport
 from ingest_validation_tools.cli_utils import ShowUsageException, exit_codes
 from ingest_validation_tools.schema_loader import PreflightError
 from ingest_validation_tools.validation_utils import (
-    get_tsv_errors, get_schema_version
+    get_tsv_errors, get_table_schema_version
 )
 
 
@@ -51,7 +51,7 @@ def main():
     try:
         schema_name = (
             args.schema if args.schema != 'metadata'
-            else get_schema_version(args.path, 'ascii').schema_name
+            else get_table_schema_version(args.path, 'ascii').schema_name
         )
     except PreflightError as e:
         errors = {'Preflight': str(e)}